pyproject.toml
[tool.poetry]
name = "pars-hh-js"
version = "0.1.0"
description = ""
authors = ["Hamanist <hamanist@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
